pyproject.toml
[project]
name = "data-ai-engineer"
version = "0.1.0"
description = "AI Engineering learning path - Phase 1: The No Magic Zone"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.27",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
]

[tool.setuptools.packages.find]
include = ["week_*"]
